Scaffold index
