--- src/Components/Publishing/__stories__/Miscellaneous.story.tsx
@@ -1,21 +1,0 @@
-import { storiesOf } from "@storybook/react"
-import React from "react"
-import { Authors, Embeds } from "../Fixtures/Components"
-import { Authors as AuthorInfo } from "../Sections/Authors"
-import { Embed } from "../Sections/Embed"
-
-storiesOf("Publishing/Miscellaneous", module)
-  .add("Embed", () => {
-    return (
-      <div style={{ width: "100%" }}>
-        <Embed section={Embeds[0]} />
-      </div>
-    )
-  })
-  .add("Author Info", () => {
-    return (
-      <div style={{ width: "100%", position: "relative" }}>
-        <AuthorInfo authors={Authors} />
-      </div>
-    )
-  })
